% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tabstatr.R
\docType{package}
\name{tabstatr}
\alias{tabstatr}
\title{tabstatr: Summary statistic tables for weighted data}
\description{
Creates tables of summary statistics for weighted data. Inspired by the Stata tabstat
command.
}
